+++ src/classes/convert/convert.php
@@ -163,6 +163,9 @@ class convert
 
 
     public function cleanup(){
+        if (!file_exists($this->intermediate_svg_file)){
+            return;
+        }
         unlink($this->intermediate_svg_file);
     }
 
@@ -174,12 +177,14 @@ class convert
     
     private function set_target_filepath()
     {
-        $this->target_filepath = ABSPATH . $this->filepath['dirname']. '/' . $this->filepath['filename'] . $this->file_suffix . '.' . $this->save_type;
+        $this->target_filepath = $this->upload_dir . '/' . $this->filepath['filename'] . $this->file_suffix . '.' . $this->save_type;
+        // $this->target_filepath = ABSPATH . $this->filepath['dirname']. '/' . $this->filepath['filename'] . $this->file_suffix . '.' . $this->save_type;
     }
     
     private function set_intermediate_svg_filepath()
     {
-        $this->intermediate_svg_file = ABSPATH . $this->filepath['dirname']. '/' . $this->filepath['filename'] . $this->file_suffix . '_intermediate.svg';
+        $this->intermediate_svg_file = $this->upload_dir . '/' . $this->filepath['filename'] . $this->file_suffix . '_intermediate.svg';
+        // $this->intermediate_svg_file = ABSPATH . $this->filepath['dirname']. '/' . $this->filepath['filename'] . $this->file_suffix . '_intermediate.svg';
     }
 
     private function rewrite_SVG_paths()
@@ -193,7 +198,7 @@ class convert
         $this->svg_data = str_replace($site_url, '/wp-content', $this->svg_data);
 
         // /wp-content
-        $this->svg_data = str_replace('/wp-content', '../../../../wp-content', $this->svg_data);
+        $this->svg_data = str_replace('/wp-content', '/../../../../wp-content', $this->svg_data);
         
     }
